+++ lib/data_migrater.rb
@@ -1,9 +1,12 @@
+# frozen_string_literal: true
+
 module DataMigrater
 end
 
 require 'data_migration'
 
 require 'data_migrater/collection'
+require 'data_migrater/csv'
 require 'data_migrater/logger'
 require 'data_migrater/migration'
 require 'data_migrater/migrator'