+++ lib/data_migration.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class DataMigration < ActiveRecord::Base
   validates :version, uniqueness: true
 end