+++ lib/generators/data_migrater/templates/config/initializers/data_migrater.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'data_migrater'
 
 DataMigrater::Migrator.new.migrate unless Rails.env.test?