+++ spec/data_migration_spec.rb
@@ -1,15 +1,17 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe DataMigration do
   let(:version) { '19841023113000' }
 
-  it { expect(DataMigration.new version: version).to be_valid }
+  it { expect(described_class.new(version: version)).to be_valid }
 
   it 'validates uniqueness' do
-    DataMigration.create! version: version
+    described_class.create! version: version
 
-    expect {
-      DataMigration.create! version: version
-    }.to raise_error ActiveRecord::RecordInvalid
+    expect do
+      described_class.create! version: version
+    end.to raise_error ActiveRecord::RecordInvalid
   end
 end