+++ spec/logger/data_logger_spec.rb
@@ -1,32 +1,30 @@
-require "spec_helper"
+# frozen_string_literal: true
 
-RSpec.describe "#data_logger" do
-  context "with :path" do
+require 'spec_helper'
+
+RSpec.describe '#data_logger' do
+  context 'with :path' do
     before do
-      stub_const "Dummy", Class.new
+      stub_const 'Dummy', Class.new
 
       Dummy.class_eval { include DataMigrater::Logger }
-      Dummy.class_eval { data_logger path: "custom.log" }
-    end
-
-    after :suite do
-      delete :dummy
+      Dummy.class_eval { data_logger path: 'custom.log' }
     end
 
     subject { Dummy.new }
 
-    it "logs on the given path file with right content" do
-      subject.logger.info "done!"
+    it 'logs on the given path file with right content' do
+      subject.logger.info 'done!'
 
-      result = /\[\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} -\d{4}\] INFO Dummy: done!\n/
+      result = /\[\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} (-|\+)\d{4}\] INFO Dummy: done!\n/
 
-      expect(File.readlines("custom.log").last).to match result
+      expect(File.readlines('custom.log').last).to match result
     end
   end
 
-  context "with no :path" do
+  context 'with no :path' do
     before do
-      stub_const "Dummy", Class.new
+      stub_const 'Dummy', Class.new
 
       Dummy.class_eval { include DataMigrater::Logger }
       Dummy.class_eval { data_logger }
@@ -34,30 +32,30 @@ RSpec.describe "#data_logger" do
 
     subject { Dummy.new }
 
-    it "logs on log folder with class name with right content" do
-      subject.logger.info "done!"
+    it 'logs on log folder with class name with right content' do
+      subject.logger.info 'done!'
 
-      result = /\[\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} -\d{4}\] INFO Dummy: done!\n/
+      result = /\[\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} (-|\+)\d{4}\] INFO Dummy: done!\n/
 
-      expect(File.readlines("log/dummy.log").last).to match result
+      expect(File.readlines('log/dummy.log').last).to match result
     end
   end
 
-  context "with no :data_logger" do
+  context 'with no :data_logger' do
     before do
-      stub_const "Dummy", Class.new
+      stub_const 'Dummy', Class.new
 
       Dummy.class_eval { include DataMigrater::Logger }
     end
 
     subject { Dummy.new }
 
-    it "logs on log folder with class name with right content" do
-      subject.logger.info "done!"
+    it 'logs on log folder with class name with right content' do
+      subject.logger.info 'done!'
 
-      result = /\[\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} -\d{4}\] INFO Dummy: done!\n/
+      result = /\[\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} (-|\+)\d{4}\] INFO Dummy: done!\n/
 
-      expect(File.readlines("log/dummy.log").last).to match result
+      expect(File.readlines('log/dummy.log').last).to match result
     end
   end
 end