+++ spec/support/methods.rb
@@ -1,3 +1,0 @@
-def delete(file_name)
-  File.delete File.expand_path("../#{file_name}.log", __dir__)
-end