+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.3.0
+
+- Added support for CSV;
+- Added options `dir` to specify the directory on Logger;
+- Added options `file` to specify the Logger file name.
+
 # 0.2.0
 
 - Added support for custom logger.